Add warehouse, not fully calibrated
There are a few errors in the OpenStudio_Standards json. The electric
equipment schedule is wrong for two of the zones.

Need to figure out how to deal with multiple thermostats, loads, etc.
as in a different one per zone.

The heating energy is off by a lot I think due to the lack of
ZoneVentilation. This object is not in OpenStudio.

The prototype buildings have a high temp radiant system but it is set
to always off. So I don’t know if we need to include it or not.

This is the same for the WaterHeater. It is there but there are no
WaterUse:Connections in the model…