So I have added a routine that runs the sizing run in it’s own folder for each scenario. This makes sure that there are no file locking issues or overwriting files.
Here are the issues that need to be resolved.

System #1
 mua_types = false will fail. (PTAC Issue Kamel Mentioned)
Control zone for SZ systems.

System #2
Sizing Convergence Errors when mua_cooling_types = DX

System #6
•	Set_vav_damper_action for NECB (Needs to be review if this is appropriate for NECB)
•	Hot water coil logic was added to prevent HW to be used when electric. (Please review if I broke anything)
•	Sizing run error when Chiller type is "Centrifugal","Rotary Screw","Reciprocating"]

System #7
Sizing Convergence Errors when mua_cooling_types = DX

I’ve commented the selections that break the tests out for now in the tester, once it is fixed just add it to the array.